chore: refactor ssh listener server (#1548)
* chore: refactor ssh listener server

I found that the ssh server in jimm could be improved in some ways:
- Use an in-memory buffered connection instead of listening on a system port. Similar to our Juju tests.
- The internal ssh package was importing internal/jimm when it shouldn't.
- Refactored the sshManager interface so that dialing the Juju controller's SSH server can be mocked.
- Moved the limitListener to a separate package.
- Removed unused functions thanks to the changes above.

* chore: Export DialControllerSSHServer again

The DialControllerSSHServer is exported again so that the DialControllerSSHServer is not doing two things in one. Arguably the SSH handler is doing too much and all the logic after the authorisation checks should be moved to a domain server. But that is a larger and separate piece of work.

* chore: fix mocks and docstrings