feat: improved integration testing (#1648)
* feat: add test for model migration to JAAS

* chore: update juju dependency

* ci: run integration tests via script

* ci: tweak integration test workflow

Our integration test workflow was a bit complicated because it allowed integration tests to be run manually or triggered via pull-request.
We had 2 steps that were conditional on how the test was triggered for starting either the development version of JIMM or using a tagged release.
Now we need to get an output from the step that starts JIMM, and to do that the step needs an ID which must be unique for each step.
Because we've never really used the manual run option, I've deleted that path and if we want to re-add it in the future, we can add it as a separate workflow.

* ci: make JIMM controller name configurable

* chore: update juju version string

* chore: add test timing

* ci: debug test

* chore: remove model migration test

The model migration test was failing in CI because our integration test uses the snapstore's latest JAAS snap plugin. In this PR we are making a change to the plugin which is needed for the test to pass.

While we could build the jaas plugin from source, I would consider that in a future change.

* chore: make ssh test executable