feat: automatic completion of internal model migration (#1670)
* feat: automatic completion of internal model migration

* refactor: use controller alias over controller tag

When Juju returns the redirect error, it doesn't return the controller tag that was provided but it does return the controller alias.
Because controller names are unique in JIMM, we can provide Juju with the controller alias and use it once the migration succeeds.

* feat: replace SetModelMigrationMode with GetModelForUpdate

The SetModelMigrationMode method was placing logic in the DB layer and after some comments on the first PR I realised that i was introducing inconsistent behavior where some DB logic was captured within the method and other logic was captured on the dbmodel struct.
This change introduces a GetModelForUpdate method that will return the model and lock the row for updates, allowing the caller to handle the business logic while adding some helper methods on the dbmodel.Model struct.

* refactor: bubble up errors in model poller

Also did minor renaming.

* feat: add ForUpdate() DB method

* chore: update camelCase to bbq-case for log keys

* chore: rename dbmodel methods