feat: handle model migrations from the model proxy (#1692)
* feat: complete internal model migration via model proxy

* Revert "feat: complete internal model migration via model proxy"

This reverts commit 7f65055eebe7c67fc8a1beb6d0dead76dab8f23e.

* feat: handle migrations from model proxy

In the model proxy it we will return an error when a model has been migrated to inform the client to try again later.
We opt for this approach rather than completing the model migration because the model proxy starts to become over-complicated when it is doing more than just proxy logic unless absolutely necessary.

* test: add tests for model migration in model-proxy