feat: make jobs facade, manager APIs and database more generic (#1711)
Rename the various bootstrap job APIs to be generic and applicable to other cases of running a juju cli command in a JIMM facade API triggered from the jaas CLI.

Starting a job is not generic, so there's GetJobInfo and StopJob, but StartBootstrapJob. I will be adding StartDestroyControllerJob in a future PR.

The manager is also not generic, under the assumption that "job" is generally bootstrap-related or similar.