fix: don't intercept ssh key methods (#1718)
* fix: don't intercept ssh key methods

The model-proxy was intercepting ssh key requests to work correctly with the ssh-proxy feature that was previously being developed.
Without that feature complete in Juju we were unintentionally preventing the addition of SSH keys to models.
In Juju 4 SSH key methods are not model-scoped so they can be implemented without intercepting the model proxy.

* chore: remove additional unneeded test