fix(batch-openfga-req): batch openfga requests to make sure we don't try to add more then the max (#1732)
openfga has a maxTuplesPerWrite config value (default to 100). We want to support JIMM.AddRelation
with more than the config value, so we are batching requests to OpenFga. This has some challenges:

 - we need the openfga charm to share this config value with us
 - we need to handle partial
failures. Both are working in progress because we are waiting on other projects' work, but this is a
good intermediate step to solve the issue to avoid the error.

partially fix #1729

Co-authored-by: Kian Parvin <46668016+kian99@users.noreply.github.com>