feat: PrepadeUpgradeTo call (#1740)
* feat: add two new manager methods to prepare and migrate & upgrade

Adds two new manager methods to handle the migrate and upgrade of a model. Very much a WIP right
now.

* feat: move prepare and migrate calls over to upgrade manager

* feat: expose Dial via JujuManager

Exposing dial via the juju manager gives us a singular place to reliably dial controllers. It is
reasonable to expect the juju manager is capable of dialing controllers, hence my choice not to
re-embed the dialer.

* feat: expose GetControllerByName

Exposes get controller by name as to not duplicate logic.

* feat: plugin exposed methods

* test: add failure case for prepare in version validation

* refactor: revert dial

* refactor: plugin dialer and store

* test: prepareupgradeto happy path test

Tests the happy path for the prepare call.

* refactor: split the migrate and upgrade call off

As the PR is growing, I've decided to do the migrate and upgrade call in another PR.

* refactor: revert permission and getControllerByName public exports

* refactor: revert facade code

* docs: godoc manager interfaces

* refactor: slim api mock

* refactor: address PR comments

* test: fix import

* refactor: go back to juju.dialer