feat: change add model logic to include target-controller  (#1745)
* chore: remove unused methods

* feat: incorporate target-controller field into model selection

Tests have been slightly tweaked and are passing.

* test: add tests for targetting controllers

Added an early exit and helpful error messages for scenarios where a user may not have access to any controllers.

* test: ensure users in tests have necessary add-model permissions

* ci: print container logs on integration test failure

* revert: keep grant jimm admin

* chore: docstring tweaks, test shuffle controllers, update Juju dep

* refactor: use domain types only in domain layer

In `internal/jimm/juju/model.go` we provided a function to convert from Juju's RPC types to our internal type. With the introduction of TargetController as a JAAS specific field we can see that this breaks down and the API types should not be used at all in the domain layer.
This commit moves the helper function up to the API layer and ensures tests use the domain type.
We also revert previous changes to the Juju import.

* chore: ensure fixture users have add-model access