fix(import-model): fix for import-model where juju does not return model's cloud region
In case of a k8s cluster Juju sometimes does not return a cloud region for a model (as there is only
one default region) and we need to set a cloud region on a model when inserting it. To that end if
and only if there is one region on the cloud (default) we select that one.